removed maxPendingBindingRequests from agent.go
added a test in agent_test.go

Note: Agent.handleInboundBindingSuccess now runs two for-loops over the
same list, which is not ideal. The benefit is that the functionality for
invalidating pending requests and identifying the valid request is
separated. Let me know if you want to move the invalidation logic into
the handleInboundBindingSuccess method.