Remove nominationRequestCount
Before then nominating a candidate pair we limited the
amount of requests we sent, and would set the connection
to failed.

This change removes that limit, and instead uses the Disconnected+Failed
timer to control this. I found this when debugging an unrelated thing,
and was surprised by the behavior. I didn't find anything in the spec
that mentioned this, and isn't a knob that users will understand how to
control easily.