Escape spaces in IPv6 ZoneID with %20
Link-local IPv6 addresses may have ZoneID attached at the end. ZoneId
is name of corresponding network interface. On Windows usually it is
"Local connection". Space in ZoneID causes problem when ICE candidate
is parsed so it has to be escaped. It looks that there is no escaping
defined specifically for SDP; the closest one is RFC 6874 for URI.
This fix partially uses this approach and replaces spaces with %20.